Read the temporal instants array with a pointer-sized stride (#38)
* Remove the accidentally-committed local Maven repository

The .m2local per-worktree Maven repository was committed by mistake. Untrack it
and ignore the directory so builds keep their local repository out of the tree.

* Read the temporal instants array with a pointer-sized stride

Temporal.instants() reads the TInstant* array returned by temporal_instants
with getPointer at Long.BYTES stride, so element i addresses a full 8-byte
pointer, matching segments(). InstantsAccessorTest asserts the instants of
[1@2019-09-01, 2@2019-09-02, 3@2019-09-03] in order; the jmeos-core suite
runs in CI.

+++ jmeos-core/src/main/java/types/temporal/Temporal.java
@@ -476,7 +476,7 @@ public abstract class Temporal<V extends Serializable> implements Serializable, 
         Pointer array= functions.temporal_instants(this.inner, intPointer);
         List<Temporal> instantList= new ArrayList<Temporal>();
         for(int i=0; i<this.num_instants(); i++){
-            Pointer p= array.getPointer((long) i *Integer.BYTES);
+            Pointer p= array.getPointer((long) i *Long.BYTES);
             Temporal t= Factory.create_temporal(p, this.getCustomType(),this.getTemporalType());
             instantList.add(t);
         }
